+++ src/parser/inline/autolink.ts
@@ -2,11 +2,10 @@
 import { combine } from '../../combinator';
 import { url } from './autolink/url';
 import { account } from './autolink/account';
+import { hashtag } from './autolink/hashtag';
 
-export const autolink: AutolinkParser = combine<HTMLAnchorElement | HTMLImageElement | HTMLSpanElement | Text, [
-  AutolinkParser.UrlParser,
-  AutolinkParser.AccountParser
-]>([
+export const autolink: AutolinkParser = combine<HTMLAnchorElement | HTMLImageElement | HTMLSpanElement | Text, AutolinkParser.InnerParsers>([
   url,
   account,
+  hashtag,
 ]);